+++ packages/stream_core_flutter/lib/src/theme.dart
@@ -27,8 +27,8 @@ export 'theme/components/stream_progress_bar_theme.dart';
 export 'theme/components/stream_reaction_picker_theme.dart';
 export 'theme/components/stream_reactions_theme.dart';
 export 'theme/components/stream_stepper_theme.dart';
+export 'theme/components/stream_switch_theme.dart';
 export 'theme/components/stream_text_input_theme.dart';
-export 'theme/components/stream_toggle_switch_theme.dart';
 
 export 'theme/primitives/stream_colors.dart';
 export 'theme/primitives/stream_icons.dart';