+++ packages/stream_core_flutter/lib/src/theme/stream_theme_extensions.dart
@@ -21,8 +21,8 @@ import 'components/stream_reaction_picker_theme.dart';
 import 'components/stream_reactions_theme.dart';
 import 'components/stream_skeleton_loading_theme.dart';
 import 'components/stream_stepper_theme.dart';
+import 'components/stream_switch_theme.dart';
 import 'components/stream_text_input_theme.dart';
-import 'components/stream_toggle_switch_theme.dart';
 import 'primitives/stream_icons.dart';
 import 'primitives/stream_radius.dart';
 import 'primitives/stream_spacing.dart';
@@ -143,6 +143,6 @@ extension StreamThemeExtension on BuildContext {
   /// Returns the [StreamStepperThemeData] from the nearest ancestor.
   StreamStepperThemeData get streamStepperTheme => StreamStepperTheme.of(this);
 
-  /// Returns the [StreamToggleSwitchThemeData] from the nearest ancestor.
-  StreamToggleSwitchThemeData get streamToggleSwitchTheme => StreamToggleSwitchTheme.of(this);
+  /// Returns the [StreamSwitchThemeData] from the nearest ancestor.
+  StreamSwitchThemeData get streamSwitchTheme => StreamSwitchTheme.of(this);
 }